Complaints to NYC Council Constituent Services

Complaints data from New York City;s CouncilStat

Usage

nycomplaints

Format

A data frame with 240,027 rows and 11 variables:

unique_key

character Complaint identifier.

account

character NYC Council District that created the casework, this may or not be the district that the constituent lives in.

opendate

date Date the casework is created by staff.

closedate

date Close Date is the date the casework was closed. A casework could be closed because the issue is resolved or because no further action could be taken.

complaint_type

character Broader category designated by staff for the constituent casework. complaint_type corresponds to NYC Council Legislative Committees.

descriptor

character Specific category designated by staff for the constituent casework.

zip

integer Zip code of the constituent.

borough

character Borough of the constituent.

city

character City of the constituent.

council_dist

Council District of the constituent.

community_board

Community Board of the constituent.

Details

Table: Data summary

Name	nycomplaints

Number of rows	240027

Number of columns	11

_______________________

Column type frequency:

Date	2

character	9

________________________

Group variables	None

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

opendate	0	1.00	2015-01-01	2022-04-07	2017-10-02	2596

closedate	16046	0.93	2015-01-02	2022-04-07	2018-01-29	2107

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

unique_key	0	1.00	12	12	0	239958	0

account	0	1.00	6	6	0	51	0

complaint_type	5060	0.98	3	25	0	56	0

descriptor	13257	0.94	2	56	0	1002	0

zip	0	1.00	5	5	0	1046	0

borough	16158	0.93	5	13	0	5	0

city	989	1.00	1	36	0	992	0

council_dist	9728	0.96	4	12	0	154	0

community_board	16091	0.93	1	16	0	94	0

Produced from NYC Open Data Initiative data. The dataset comes from CouncilStat, which is used by many NYC Council district offices to enter and track constituent cases that can range from issues around affordable housing, to potholes and pedestrian safety. This dataset aggregates the information that individual staff have input. However, district staffs handle a wide range of complex issues. Each offices uses the program differently, and thus records cases, differently and so comparisons between accounts may be difficult. Not all offices use the program. For more info - http://labs.council.nyc/districts/data/

References

https://data.cityofnewyork.us/City-Government/NYC-Council-Constituent-Services/9hzi-kbqb,http://labs.council.nyc/districts/data/

Author

Kieran Healy